# Apply small optimizations to HTTPPropagator.extract