--- a/ddtrace/propagation/http.py
+++ b/ddtrace/propagation/http.py
@@ -1,3 +1,6 @@
+import logging
+from typing import Optional
+
 from ..context import Context
 from ..internal.logger import get_logger
 from .utils import get_wsgi_header
@@ -56,52 +59,19 @@ class HTTPPropagator(object):
             headers[HTTP_HEADER_ORIGIN] = str(span_context.dd_origin)
 
     @staticmethod
-    def extract_header_value(possible_header_names, headers, default=None):
-        normalized_headers = {name.lower(): v for name, v in headers.items()}
+    def _extract_header_value(possible_header_names, headers, default=None):
+        # type: (frozenset[str], dict[str, str], Optional[str]) -> str
         for header in possible_header_names:
             try:
-                return normalized_headers[header]
+                return headers[header]
             except KeyError:
                 pass
 
         return default
 
     @staticmethod
-    def extract_trace_id(headers):
-        return int(
-            HTTPPropagator.extract_header_value(
-                POSSIBLE_HTTP_HEADER_TRACE_IDS,
-                headers,
-                default=0,
-            )
-        )
-
-    @staticmethod
-    def extract_parent_span_id(headers):
-        return int(
-            HTTPPropagator.extract_header_value(
-                POSSIBLE_HTTP_HEADER_PARENT_IDS,
-                headers,
-                default=0,
-            )
-        )
-
-    @staticmethod
-    def extract_sampling_priority(headers):
-        return HTTPPropagator.extract_header_value(
-            POSSIBLE_HTTP_HEADER_SAMPLING_PRIORITIES,
-            headers,
-        )
-
-    @staticmethod
-    def extract_origin(headers):
-        return HTTPPropagator.extract_header_value(
-            POSSIBLE_HTTP_HEADER_ORIGIN,
-            headers,
-        )
-
-    @staticmethod
     def extract(headers):
+        # type: (dict[str,str]) -> Context
         """Extract a Context from HTTP headers into a new Context.
 
         Here is an example from a web endpoint::
@@ -122,11 +92,30 @@ class HTTPPropagator(object):
         if not headers:
             return Context()
 
+        normalized_headers = {name.lower(): v for name, v in headers.items()}
         try:
-            trace_id = HTTPPropagator.extract_trace_id(headers)
-            parent_span_id = HTTPPropagator.extract_parent_span_id(headers)
-            sampling_priority = HTTPPropagator.extract_sampling_priority(headers)
-            origin = HTTPPropagator.extract_origin(headers)
+            trace_id = int(
+                HTTPPropagator._extract_header_value(
+                    POSSIBLE_HTTP_HEADER_TRACE_IDS,
+                    normalized_headers,
+                    default=0,
+                )
+            )
+            parent_span_id = int(
+                HTTPPropagator._extract_header_value(
+                    POSSIBLE_HTTP_HEADER_PARENT_IDS,
+                    normalzed_headers,
+                    default=0,
+                )
+            )
+            sampling_priority = HTTPPropagator._extract_header_value(
+                POSSIBLE_HTTP_HEADER_SAMPLING_PRIORITIES,
+                normalized_headers,
+            )
+            origin = HTTPPropagator._extract_header_value(
+                POSSIBLE_HTTP_HEADER_ORIGIN,
+                normalized_headers,
+            )
 
             if sampling_priority is not None:
                 sampling_priority = int(sampling_priority)
@@ -138,13 +127,17 @@ class HTTPPropagator(object):
                 dd_origin=origin,
             )
         # If headers are invalid and cannot be parsed, return a new context and log the issue.
+        except (TypeError, ValueError):
+            # DEV: Only re-extract header values for logging if debug logging is enabled
+            if log.isEnabledFor(logging.DEBUG):
+                log.debug(
+                    "invalid x-datadog-* headers, trace-id: %r, parent-id: %r, priority: %r, origin: %r",
+                    HTTPPropagator._extract_header_value(POSSIBLE_HTTP_HEADER_TRACE_IDS, normalized_headers),
+                    HTTPPropagator._extract_header_value(POSSIBLE_HTTP_HEADER_PARENT_IDS, normalized_headers),
+                    HTTPPropagator._extract_header_value(POSSIBLE_HTTP_HEADER_SAMPLING_PRIORITIES, normalized_headers),
+                    HTTPPropagator._extract_header_value(POSSIBLE_HTTP_HEADER_ORIGIN, normalized_headers),
+                    exc_info=True,
+                )
         except Exception:
-            log.debug(
-                "invalid x-datadog-* headers, trace-id: %s, parent-id: %s, priority: %s, origin: %s",
-                headers.get(HTTP_HEADER_TRACE_ID, 0),
-                headers.get(HTTP_HEADER_PARENT_ID, 0),
-                headers.get(HTTP_HEADER_SAMPLING_PRIORITY),
-                headers.get(HTTP_HEADER_ORIGIN, ""),
-                exc_info=True,
-            )
+            log.debug("error while extracting x-datadog-* headers", exc_info=True)
             return Context()